+++ backend/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Boolean, ForeignKey, Integer, String
+from sqlalchemy import Boolean, ForeignKey, Integer, Numeric, String
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -21,7 +21,7 @@ class UserProfile(Base, IdMixin, TimestampMixin):
     __tablename__ = "user_profiles"
 
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), unique=True)
-    gold: Mapped[int] = mapped_column(BigInteger, default=0)
+    gold: Mapped[float] = mapped_column(Numeric(40, 0), default=0)
     farm_atk_bonus: Mapped[float] = mapped_column(default=0)
     add_attack_speed: Mapped[float] = mapped_column(default=0)
     current_character_id: Mapped[str] = mapped_column(String(80), default="weapon_master")
